requirements.txt
django
djangorestframework
celery 
rabbitmq-server
kombu